Move tags handling into jobs-moderation
